Move rest of Encodable/Decodable impls
As we did for the primitive and string types, put implementations of
`Encodable` and `Decodable` in the `impls` module.
